test: preserve Python interpreter fallback
