+++ crates/rumoca/tests/backend_template_runtime_regression.rs
@@ -18,6 +18,12 @@ use rumoca_phase_codegen::templates;
 use rumoca_sim::{SimOptions, SimResult, SimSolverMode, simulate_dae_with_diagnostics};
 use tempfile::Builder;
 
+#[cfg(feature = "template-runtime-tests")]
+#[path = "backend_template_runtime_regression/python_runtime.rs"]
+mod python_runtime;
+#[cfg(feature = "template-runtime-tests")]
+use python_runtime::run_python;
+
 // ============================================================================
 // Tolerance — max bounded relative error: |a-b| / max(|a|, |b|, 1.0)
 // ============================================================================
@@ -29,121 +35,6 @@ const CASADI_TOLERANCE: f64 = 0.01;
 /// Embedded C uses RK4 with dt=0.001, FMI2 uses forward Euler with dt=0.001.
 #[cfg(feature = "template-runtime-tests")]
 const C_TOLERANCE: f64 = 0.05;
-
-// ============================================================================
-// Runtime detection
-// ============================================================================
-
-#[cfg(feature = "template-runtime-tests")]
-fn python_modules(backend: &str) -> &'static [&'static str] {
-    match backend {
-        "CasADi" => &["casadi", "numpy"],
-        "SymPy" => &["sympy"],
-        "ONNX" => &["onnx", "onnxruntime", "numpy"],
-        "JAX" => &["jax", "diffrax", "numpy"],
-        _ => panic!("unknown Python template backend: {backend}"),
-    }
-}
-
-#[cfg(feature = "template-runtime-tests")]
-fn discover_python<'a>(candidates: impl IntoIterator<Item = &'a str>) -> Option<&'a str> {
-    candidates.into_iter().find(|candidate| {
-        Command::new(candidate)
-            .arg("--version")
-            .output()
-            .is_ok_and(|output| output.status.success())
-    })
-}
-
-#[cfg(feature = "template-runtime-tests")]
-fn python_command() -> &'static str {
-    discover_python(["python3", "python"])
-        .unwrap_or_else(|| panic!("expected python3 or python to be available"))
-}
-
-#[cfg(feature = "template-runtime-tests")]
-fn probe_python_modules(interpreter: &str, backend: &str, modules: &[&str]) -> Result<(), String> {
-    let output = Command::new(interpreter)
-        .args(["-c", &format!("import {}", modules.join(", "))])
-        .output()
-        .map_err(|error| {
-            format!("{backend} dependency probe could not run {interpreter}: {error}")
-        })?;
-    if output.status.success() {
-        return Ok(());
-    }
-    Err(format!(
-        "{backend} Python runtime dependency probe failed\ninterpreter: {interpreter}\nrequired modules: {}\nstdout:\n{}\nstderr:\n{}",
-        modules.join(", "),
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr),
-    ))
-}
-
-#[cfg(feature = "template-runtime-tests")]
-fn require_python_modules(backend: &str, modules: &[&str]) {
-    if let Err(error) = probe_python_modules(python_command(), backend, modules) {
-        panic!("{error}");
-    }
-}
-
-#[cfg(all(feature = "template-runtime-tests", unix))]
-fn probe_executable(dir: &std::path::Path, name: &str, body: &str) -> std::path::PathBuf {
-    use std::os::unix::fs::PermissionsExt;
-
-    let path = dir.join(name);
-    fs::write(&path, body).expect("write fake Python executable");
-    let mut permissions = fs::metadata(&path)
-        .expect("fake Python metadata")
-        .permissions();
-    permissions.set_mode(0o755);
-    fs::set_permissions(&path, permissions).expect("make fake Python executable");
-    path
-}
-
-#[test]
-#[cfg(feature = "template-runtime-tests")]
-fn python_runtime_probe_contract_modules_are_backend_specific() {
-    assert_eq!(python_modules("CasADi"), &["casadi", "numpy"]);
-    assert_eq!(python_modules("SymPy"), &["sympy"]);
-    assert_eq!(python_modules("ONNX"), &["onnx", "onnxruntime", "numpy"]);
-    assert_eq!(python_modules("JAX"), &["jax", "diffrax", "numpy"]);
-}
-
-#[test]
-#[cfg(all(feature = "template-runtime-tests", unix))]
-fn python_runtime_probe_contract_preserves_backend_import_errors() {
-    let dir = Builder::new()
-        .prefix("rumoca_python_probe_")
-        .tempdir()
-        .expect("create Python probe dir");
-    let args_log = dir.path().join("args");
-    let executable = probe_executable(
-        dir.path(),
-        "python-isolated",
-        &format!(
-            "#!/bin/sh\nif [ \"$1\" = --version ]; then printf '%s\\n' \"$*\" > \"{}\"; exit 0; fi\necho \"ModuleNotFoundError: No module named 'onnxruntime'\" >&2\nexit 7\n",
-            args_log.display()
-        ),
-    );
-
-    let candidate = executable.to_str().expect("UTF-8 executable path");
-    assert_eq!(discover_python([candidate]), Some(candidate));
-    assert_eq!(
-        fs::read_to_string(args_log).expect("read probe args"),
-        "--version\n"
-    );
-
-    let error = probe_python_modules(candidate, "ONNX", python_modules("ONNX"))
-        .expect_err("missing ONNX dependency must fail closed");
-    for expected in [
-        "ONNX",
-        "onnx, onnxruntime, numpy",
-        "ModuleNotFoundError: No module named 'onnxruntime'",
-    ] {
-        assert!(error.contains(expected), "missing {expected:?}: {error}");
-    }
-}
 
 #[cfg(feature = "template-runtime-tests")]
 fn strict_runtime_dependencies() -> bool {
@@ -391,37 +282,6 @@ fn assert_traces_match(
             "{backend_name}: state '{name}' deviation {dev:.4e} exceeds tolerance {tolerance:.4e}"
         );
     }
-}
-
-// ============================================================================
-// Python execution helper (returns stdout)
-// ============================================================================
-
-#[cfg(feature = "template-runtime-tests")]
-fn run_python(rendered: &str, driver: &str, backend: &str) -> String {
-    require_python_modules(backend, python_modules(backend));
-    let dir = Builder::new()
-        .prefix("rumoca_runtime_test_")
-        .tempdir()
-        .expect("create temp dir");
-    let model_path = dir.path().join("model.py");
-    let driver_path = dir.path().join("driver.py");
-    fs::write(&model_path, rendered).expect("write model.py");
-    fs::write(&driver_path, driver).expect("write driver.py");
-
-    let output = Command::new(python_command())
-        .arg(driver_path.to_str().unwrap())
-        .output()
-        .expect("run Python driver");
-
-    assert!(
-        output.status.success(),
-        "Python execution failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    String::from_utf8(output.stdout).expect("stdout is utf8")
 }
 
 #[cfg(feature = "template-runtime-tests")]
